Start to fix parsing of RESP
Having realised how wrong I got the initial parsing of RESP thanks to
mis-understanding of REDIS documentation, this starts work on reading
lines and parsing arguments etc. to get valid Resp commands from the
incoming connection.

Adds a generator function to produce a line reader from an io.Reader (to
which a connection can be cast - starting to see the power of
non-explicit interfaces!). From this need to finish generator function
to produce resp commands from the incoming connection.