Add very basic TTL / expiry for keys
Addition makes it obvious that some refactoring is required to add this
onto the data source, and as a separate structure to handle timing /
expiry. Also makes it clear some refactoring is required for handlers so
not all work is done within server.go which is already getting too long.

Finally, to support some of the smarter functions of Redis e.g. blocking
clients etc., the router.Responder is not adequate and will have to be
changed.

The next piece of work should really be refactoring to a point where all
functionality of Redis can be added without the existing code getting
messy.