Begin refactoring to make more extensible
Begin by changing router to no longer use 'Responder' concept which was
poorly named to begin with, and instead to use a Connection interface
specified in its own package and implemented by the client connection
struct. This in turn gives back a response writer which is implemented
by the existing struct but described by a new interface.